延迟任务质量加固 Signed-off-by: hxf <haoxiaofeng2@huawei.com>

+++ services/native/include/work_policy_manager.h
@@ -292,7 +292,7 @@ private:
         const std::string &abilityName);
     void SendIdeWorkRetriggerEvent(int32_t delaytime);
     bool IsSpecialScene(std::shared_ptr<WorkStatus> topWork);
-    long long GetCurrentTimeMs();
+    uint64_t GetCurrentTimeMs();
     std::string GetConditionString(const std::shared_ptr<WorkStatus> workStatus);
 
 private:
